jlfblog/requirements.txt
alembic==1.0.5
aniso8601==4.0.1
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
colorama==0.4.1
cryptography==2.3.1
Django==2.0.6
django-redis==4.10.0
django-redis-cache==1.8.1
djangorestframework==3.9.0
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Migrate==2.3.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mysqlclient==1.3.13
nltk==3.4
Pillow==5.3.0
pycparser==2.19
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.8.0
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
qrcode==6.1
redis==2.10.6
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
